some MSVC-related fixes, thanks to Laurence Muller

+++ src/drl_graph.cpp
@@ -1142,7 +1142,7 @@ void graph::Solve_Analytic( int node_ind, float &pos_x, float &pos_y )
    // Don't cut at end of scale
    if ( CUT_END >= 39500 ) return;
 
-   float num_connections = sqrt(neighbors[node_ind].size());
+   float num_connections = sqrt((double)neighbors[node_ind].size());
    float maxLength = 0;
 
    map<int, float>::iterator maxIndex;
